refactor: 코드 리팩토링
- Event 클래스 위치 변경

- SSE 구독 요청 경로 RequestMatcher 변경
